refactor(console): shared decompile-project core (kuna_console::project)
Move FuncResult (+proto), decompile_targets, render_c, and the four
decompile-project artifact builders (build_header/build_c/build_asm/
build_readme + collect_dat_addrs/sanitize_guard) from kuna-cli into a new
shared, wasm32-wasip1-safe kuna-console module so the kuna_wasm front-end
can reach them without kuna-cli's subprocess machinery. kuna-cli's
decompile_all.rs/decompile_project.rs become thin wrappers; build_readme
gains a path_label param (CLI passes binary_path.display(), output
byte-identical — verified: decompile-all --json and the project folder
diff empty vs pre-refactor on sample.elf and /usr/bin/true). kuna-console
promotes object from dev-dep to dep (already in the workspace graph).
