+++ decompiler/crates/kuna-console/src/lib.rs
@@ -20,6 +20,7 @@ pub mod interface;
 pub mod ifaceterm;
 pub mod ifacedecomp;
 pub mod engine;
+pub mod project;
 pub mod codedata;
 pub mod kuna_console;
 pub mod grammar;
